Make DhtArc unidirectional -- defined by left edge rather than center
This is a huge change, particularly to gossip behavior. With
bidirectional arcs, when peers have roughly equivalently sized arcs,
half of the peers who have overlapping arcs will not
see each other or gossip with each other because their centerpoints are
not contained within each others' arcs. With unidirectional arcs, this
problem is removed at the expense of making peer discovery
asymmmetrical.